Add files via upload
The attached file is a final project I did as part of my statistics degree. It's an exploratory analysis of Missouri unemployment from 2008-2012. The data can be found from data.gov (I attempted to upload it directly, but it seems to be too large a file). The name of the dataset can be found by viewing the "read.csv()" line of my code. I can also provide the data directly if requested.